v0.2.0(faceijk): widen maxDim/unitScale tables → int64_t[23] + locals
Second v0.2.0 widening per POC-5. Eliminates int32 truncation for
the ext-res table arithmetic that drives _adjustOverageClassII and
the pentagon overage walk.

Changes:
- faceijk.c:315-339 — maxDimByCIIres extended to length 23, element
  type int → int64_t. New entries: [21]=-1, [22]=2*7^11=3,954,653,486
  (the value Session 2 had to defer).
- faceijk.c:345-371 — unitScaleByCIIres extended to length 23, same
  pattern. New entries: [21]=-1, [22]=7^11=1,977,326,743.
- faceijk.c:617-622, 783-788 — pentagon overage Vec2d edge bounds:
  int maxDim → int64_t maxDim, with explicit (double)maxDim casts in
  the floating-point Vec2d initializers.
- faceijk.c:925-926 — _adjustOverageClassII: int maxDim → int64_t.
  `maxDim *= 3` no longer overflows at res 22 (3 * 3.95e9 = 1.19e10).
- faceijk.c:965-966 — _adjustOverageClassII: int unitScale → int64_t.
  `unitScale *= 3` no longer overflows at res 22 (3 * 7^11 = 5.93e9).
- coordijk.h:108-112 — _setIJK params widened int → int64_t (called
  from _adjustOverageClassII at faceijk.c:947 with maxDim arg).
- coordijk.h:155-159 — _ijkScale factor widened int → int64_t (called
  with int64_t unitScale at faceijk.c:967, 609).

Local ctest 326/326 PASS under ASAN+UBSAN. POC-5 still 85/85 (its
analytic model unchanged; the boundary it predicts is now corrected
in the actual library).

Subsequent commits: testCoordIjkExtRes regression suite (D2-G2 res
20-22 round-trip closure) + audit.py reclassification of these sites
from deferred:post-mvp to widened.

+++ src/h3lib/include/coordijk.h
@@ -105,7 +105,7 @@ typedef enum {
  * @param j The desired j component value.
  * @param k The desired k component value.
  */
-static inline void _setIJK(CoordIJK *ijk, int i, int j, int k) {
+static inline void _setIJK(CoordIJK *ijk, int64_t i, int64_t j, int64_t k) {
     ijk->i = i;
     ijk->j = j;
     ijk->k = k;
@@ -157,7 +157,7 @@ static inline void _ijkSub(const CoordIJK *h1, const CoordIJK *h2,
  * @param c The ijk coordinates to scale.
  * @param factor The scaling factor.
  */
-static inline void _ijkScale(CoordIJK *c, int factor) {
+static inline void _ijkScale(CoordIJK *c, int64_t factor) {
     c->i *= factor;
     c->j *= factor;
     c->k *= factor;
